+++ executors/keptn/pkg/keptn/consts.go
@@ -1,0 +1,17 @@
+package keptn
+
+const (
+	sliFilename = "sli.yaml"
+	sliURI      = "prometheus/sli.yaml"
+	sloFilename = "slo.yaml"
+	sloURI      = "slo.yaml"
+
+	jobExecutorFilename = "config.yaml"
+	jobExecutorURI      = "job/config.yaml"
+
+	ShipyardFileName    string = "shipyard.yaml"
+	KeptnConfigFileName string = "keptn-config.json"
+
+	KeptnAuthTokenKey     string = "x-token"
+	HTTPTransportProtocol string = "http"
+)
