+++ executors/keptn/pkg/keptn/git.go
@@ -1,0 +1,7 @@
+package keptn
+
+type Git struct {
+	URL   string
+	Token string
+	User  string
+}
