--- executors/keptn/pkg/keptn/testwith/helmrelease.yaml
@@ -1,13 +1,0 @@
-apiVersion: helm.fluxcd.io/v2beta1
-kind: HelmRelease
-metadata:
-  name: app 
-  namespace: webserver
-spec:
-  interval: 5m
-  chart:
-    spec:
-      chart: webserver
-      sourceRef:
-        kind: HelmRepository
-        name: webserver
