--- executors/keptn/pkg/keptn/testwith/hey/event.json
@@ -1,13 +1,0 @@
-{
-	"type": "sh.keptn.event.dev.continuousvalidation.triggered",
-	"specversion": "1.0",
-	"source": "test-events",
-	"contenttype": "application/json",
-	"data": {
-		"project": "hey",
-		"stage": "dev",
-		"service": "webservers"
-	}
-}
-
-
