--- executors/keptn/pkg/keptn/testwith/job-executor-service.yaml
@@ -1,168 +1,0 @@
----
-# Deployment of our job-executor-service
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: job-executor-service
-  namespace: orkestra
-spec:
-  selector:
-    matchLabels:
-      run: job-executor-service
-  replicas: 1
-  template:
-    metadata:
-      labels:
-        run: job-executor-service
-        app.kubernetes.io/name: job-executor-service
-        app.kubernetes.io/version: 0.1.3
-    spec:
-      containers:
-        - name: job-executor-service
-          image: keptnsandbox/job-executor-service:0.1.3
-          ports:
-            - containerPort: 8080
-          resources:
-            limits:
-              cpu: 1
-              memory: 512Mi
-            requests:
-              cpu: 50m
-              memory: 128Mi
-          env:
-            - name: INIT_CONTAINER_CONFIGURATION_SERVICE_API_ENDPOINT
-              value: "http://configuration-service:8080"
-            - name: CONFIGURATION_SERVICE
-              value: 'http://configuration-service:8080'
-            - name: JOB_NAMESPACE
-              value: 'orkestra'
-            - name: INIT_CONTAINER_IMAGE
-              value: 'keptnsandbox/job-executor-service-initcontainer:0.1.3'
-            - name: DEFAULT_RESOURCE_LIMITS_CPU
-              value: "1"
-            - name: DEFAULT_RESOURCE_LIMITS_MEMORY
-              value: "512Mi"
-            - name: DEFAULT_RESOURCE_REQUESTS_CPU
-              value: "50m"
-            - name: DEFAULT_RESOURCE_REQUESTS_MEMORY
-              value: "128Mi"
-        - name: distributor
-          image: keptn/distributor:0.8.4
-          livenessProbe:
-            httpGet:
-              path: /health
-              port: 10999
-            initialDelaySeconds: 5
-            periodSeconds: 5
-          imagePullPolicy: Always
-          ports:
-            - containerPort: 8080
-          resources:
-            requests:
-              memory: "32Mi"
-              cpu: "50m"
-            limits:
-              memory: "128Mi"
-              cpu: "500m"
-          env:
-            - name: PUBSUB_URL
-              value: 'nats://keptn-nats-cluster'
-            - name: PUBSUB_TOPIC
-              value: 'sh.keptn.>'
-            - name: PUBSUB_RECIPIENT
-              value: '127.0.0.1'
-            - name: VERSION
-              valueFrom:
-                fieldRef:
-                  apiVersion: v1
-                  fieldPath: 'metadata.labels[''app.kubernetes.io/version'']'
-            - name: K8S_DEPLOYMENT_NAME
-              valueFrom:
-                fieldRef:
-                  apiVersion: v1
-                  fieldPath: 'metadata.labels[''app.kubernetes.io/name'']'
-            - name: K8S_POD_NAME
-              valueFrom:
-                fieldRef:
-                  apiVersion: v1
-                  fieldPath: metadata.name
-            - name: K8S_NAMESPACE
-              valueFrom:
-                fieldRef:
-                  apiVersion: v1
-                  fieldPath: metadata.namespace
-            - name: K8S_NODE_NAME
-              valueFrom:
-                fieldRef:
-                  apiVersion: v1
-                  fieldPath: spec.nodeName
-      serviceAccountName: job-executor-service
----
-# Expose job-executor-service via Port 8080 within the cluster
-apiVersion: v1
-kind: Service
-metadata:
-  name: job-executor-service
-  namespace: orkestra
-  labels:
-    run: job-executor-service
-spec:
-  ports:
-    - port: 8080
-      protocol: TCP
-  selector:
-    run: job-executor-service
----
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: job-executor-service
-  namespace: orkestra
----
-# Role for accessing secrets in the namespace
-apiVersion: rbac.authorization.k8s.io/v1
-kind: Role
-metadata:
-  name: job-executor-service
-  namespace: orkestra
-rules:
-  - apiGroups:
-      - ""
-    resources:
-      - "secrets"
-    verbs:
-      - "get"
-      - "list"
-      - "watch"
-  - apiGroups:
-      - ""
-    resources:
-      - "pods"
-      - "pods/log"
-      - "persistentvolumeclaims"
-      - "jobs"
-    verbs:
-      - "*"
-  - apiGroups:
-      - "batch"
-      - "extensions"
-    resources:
-      - "jobs"
-    verbs:
-      - "*"
----
-# Bind role for accessing secrets onto the job-executor-service service account
-apiVersion: rbac.authorization.k8s.io/v1
-kind: RoleBinding
-metadata:
-  name: job-executor-service
-  namespace: orkestra
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: Role
-  name: job-executor-service
-subjects:
-  - kind: ServiceAccount
-    name: job-executor-service
-    namespace: orkestra
-
