--- executors/keptn/pkg/keptn/testwith/job/README.md
@@ -1,3 +1,0 @@
-# Config.yaml
-
-`config.yaml` is required for triggering `hey` load-testing using the [Keptn Job Executor Service](https://github.com/keptn-sandbox/job-executor-service).
