--- executors/keptn/pkg/keptn/testwith/keptn-config.json
@@ -1,11 +1,0 @@
-{
-  "url": "http://52.247.18.168/api",
-  "namespace": "orkestra",
-  "timeframe": "5m",
-  "token": {
-    "secretRef":  {
-      "name": "keptn-api-token",
-      "namespace": "orkestra"
-    }
-  }
-}
