--- executors/keptn/pkg/keptn/testwith/shipyard.yaml
@@ -1,14 +1,0 @@
-apiVersion: "spec.keptn.sh/0.2.2"
-kind: "Shipyard"
-metadata:
-  name: "shipyard-webserver"
-spec:
-  stages:
-    - name: "dev"
-      sequences:
-        - name: "evaluation"
-          tasks:
-          - name: "test"
-            properties:
-              teststrategy: "functional"
-          - name: "evaluation"
