--- executors/keptn/pkg/keptn/testwith/slo.yaml
@@ -1,23 +1,0 @@
-spec_version: '1.0'
-comparison:
-  compare_with: "single_result"
-  include_result_with_score: "pass"
-  aggregate_function: avg
-objectives:
-  - sli: response_time_p99
-    pass:
-      - criteria:
-          - "<0.15"
-    warning:
-      - criteria:
-          - "<=0.05"
-  - sli: error_percentage
-    pass:
-      - criteria:
-        - "<10"
-    warning:
-      - criteria:
-        - "<=5"
-total_score:
-  pass: "100%"
-  warning: "75%"
