+++ executors/keptn/testing/keptn-prometheus-service.yaml
@@ -1,0 +1,211 @@
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: keptn-prometheus-service
+  namespace: orkestra
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRole
+metadata:
+  name: keptn-create-prom-clusterrole
+rules:
+  - apiGroups:
+      - "rbac.authorization.k8s.io"
+    resources:
+      - clusterroles
+      - clusterrolebindings
+    verbs:
+      - get
+      - create
+      - update
+    resourceNames:
+      - "prometheus"
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRole
+metadata:
+  name: keptn-prom-prometheus
+rules:
+  - apiGroups:
+      - ""
+    resources:
+      - endpoints
+      - nodes
+      - nodes/proxy
+      - pods
+      - services
+    verbs:
+      - get
+      - list
+      - watch
+  - apiGroups:
+      - "extensions"
+    resources:
+      - "ingresses"
+    verbs:
+      - get
+      - list
+      - watch
+  - nonResourceURLs: [ "/metrics" ]
+    verbs: [ "get" ]
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRole
+metadata:
+  name: keptn-read-secret-prometheus
+rules:
+  - apiGroups:
+      - ""
+    resources:
+      - secrets
+    verbs:
+      - get
+      - watch
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
+metadata:
+  name: keptn-prometheus-sli-service
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: ClusterRole
+  name: keptn-read-secret-prometheus
+subjects:
+  - kind: ServiceAccount
+    name: keptn-prometheus-service
+    namespace: orkestra
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
+metadata:
+  name: keptn-create-prom-clusterrole
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: ClusterRole
+  name: keptn-create-prom-clusterrole
+subjects:
+  - kind: ServiceAccount
+    name: keptn-prometheus-service
+    namespace: orkestra
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
+metadata:
+  name: keptn-prom-prometheus
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: ClusterRole
+  name: keptn-prom-prometheus
+subjects:
+  - kind: ServiceAccount
+    name: keptn-prometheus-service
+    namespace: orkestra
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: RoleBinding
+metadata:
+  name: keptn-keptndomain-prom-service
+  namespace: orkestra
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: Role
+  name: keptn-read-keptndomain
+subjects:
+  - kind: ServiceAccount
+    name: keptn-prometheus-service
+    namespace: orkestra
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: prometheus-service
+  namespace: orkestra
+spec:
+  selector:
+    matchLabels:
+      run: prometheus-service
+  replicas: 1
+  template:
+    metadata:
+      labels:
+        run: prometheus-service
+    spec:
+      containers:
+        - name: prometheus-service
+          image: keptncontrib/prometheus-service:0.6.0
+          ports:
+            - containerPort: 8080
+          resources:
+            requests:
+              memory: "32Mi"
+              cpu: "50m"
+            limits:
+              memory: "256Mi"
+              cpu: "500m"
+          env:
+            - name: METRICS_SCRAPE_PATH
+              value: '/metrics'
+            - name: CONFIGURATION_SERVICE
+              value: 'http://configuration-service:8080'
+            - name: EVENTBROKER
+              value: 'http://localhost:8081/event'
+            - name: API
+              value: 'ws://api-service:8080/websocket'
+            - name: PROMETHEUS_NS
+              value: 'prometheus'
+            - name: PROMETHEUS_CM
+              value: 'prometheus-server'
+            - name: PROMETHEUS_LABELS
+              value: 'component=server'
+            - name: PROMETHEUS_ENDPOINT
+              value: "http://prometheus-server.prometheus.svc.cluster.local:80"
+            - name: PROMETHEUS_CONFIG_FILENAME
+              value: 'prometheus.yml'
+            - name: ALERT_MANAGER_CONFIG_FILENAME
+              value: 'alertmanager.yml'
+            - name: ALERT_MANAGER_CM
+              value: 'prometheus-alertmanager'
+            - name: ALERT_MANAGER_LABELS
+              value: 'component=alertmanager'
+            - name: ALERT_MANAGER_NS
+              value: 'prometheus'
+            - name: ALERT_MANAGER_TEMPLATE_CM
+              value: 'alertmanager-templates'
+            - name: POD_NAMESPACE
+              valueFrom:
+                fieldRef:
+                  fieldPath: metadata.namespace
+        - name: distributor
+          image: keptn/distributor:0.8.3
+          ports:
+            - containerPort: 8080
+          resources:
+            requests:
+              memory: "16Mi"
+              cpu: "25m"
+            limits:
+              memory: "128Mi"
+              cpu: "250m"
+          env:
+            - name: PUBSUB_URL
+              value: 'nats://keptn-nats-cluster'
+            - name: PUBSUB_TOPIC
+              value: 'sh.keptn.event.monitoring.configure,sh.keptn.event.get-sli.triggered'
+            - name: PUBSUB_RECIPIENT
+              value: 'prometheus-service'
+      serviceAccountName: keptn-prometheus-service
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: prometheus-service
+  namespace: orkestra
+  labels:
+    run: prometheus-service
+spec:
+  ports:
+    - port: 8080
+      protocol: TCP
+  selector:
+    run: prometheus-service
+
