+++ executors/keptn/testing/shipyard.yaml
@@ -1,0 +1,14 @@
+apiVersion: "spec.keptn.sh/0.2.2"
+kind: "Shipyard"
+metadata:
+  name: "shipyard-webserver"
+spec:
+  stages:
+    - name: "dev"
+      sequences:
+        - name: "evaluation"
+          tasks:
+          - name: "test"
+            properties:
+              teststrategy: "functional"
+          - name: "evaluation"
